Detect crashing processes
