+++ core/src/lib.rs
@@ -10,9 +10,9 @@ pub mod host;
 pub mod http;
 pub mod keystone;
 mod posix_module;
+mod posix_process;
 mod posix_spawn;
 mod proxy;
-mod spawn;
 mod sqlite;
 
 capnp_import::capnp_import!("schema/**/*.capnp");
